Remove unused files, rename some files, clean up code, Course_IDs in F4
